requirements.txt
dj-database-url==0.5.0
Django==1.11
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.6.1
pytz==2018.7
uuid==1.30
